ai-hackathon-use-cases/Inter-Agency-Knowledge-Hub/requirements.txt
# Core Azure AI
azure-ai-projects>=1.0.0
azure-ai-evaluation>=1.13.7
azure-ai-inference>=1.0.0
azure-identity>=1.19.0

# Microsoft Graph for Entra ID
msgraph-sdk>=1.0.0

# Azure AI Search
azure-search-documents>=11.4.0

# Microsoft Agent Framework
semantic-kernel>=1.45.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0

# Utilities
python-dotenv>=1.0.0
